arm64: CI Code scanning warning clean
driver inclusion
category: bugfix
bugzilla: 13683
CVE: NA

-------------------------------------------------

CI Code scanning warning clean

Feature or Bugfix:Bugfix

+++ include/linux/uacce.h
@@ -23,8 +23,8 @@ struct uacce_qfile_region {
 	unsigned long iova;	/* iova share between user and device space */
 	struct page **pages;
 	unsigned int nr_pages;
-	unsigned long prot;
-	int flags;
+	int prot;
+	unsigned int flags;
 	struct list_head qs;	/* qs sharing the same region, for ss */
 	void *kaddr;		/* kernel addr, for dko */
 	dma_addr_t dma;		/* dma address, if created by dma api */
@@ -87,7 +87,7 @@ struct uacce {
 	const char *drv_name;
 	const char *algs;
 	const char *api_ver;
-	int flags;
+	unsigned int flags;
 	unsigned long qf_pg_start[UACCE_QFRT_MAX];
 	int status;
 	struct uacce_ops *ops;
